requirements.txt
arrow==1.3.0
attrs==24.2.0
autopep8==2.3.1
beautifulsoup4==4.12.3
binaryornot==0.4.4
blessed==1.20.0
blinker==1.8.2
bs4==0.0.2
cachelib==0.9.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
DateTime==5.5
distlib==0.3.9
editor==1.6.6
fake-useragent==1.5.1
filelock==3.16.1
Flask==3.0.3
Flask-Caching==2.3.0
h11==0.14.0
idna==3.10
inquirer==3.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
langcodes==3.4.1
language_data==1.2.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
outcome==1.3.0.post0
packaging==24.1
pillow==11.0.0
pipenv==2024.1.0
pipfile-requirements==0.3.0
platformdirs==4.3.6
pycodestyle==2.12.1
pycryptodome==3.21.0
pydub==0.25.1
pyfiglet==1.0.2
Pygments==2.18.0
PyPasser==0.0.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
readchar==4.2.0
requests==2.32.3
rich==13.9.2
runs==1.2.2
selenium==4.25.0
setuptools==75.1.0
sh==2.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.8.1
text-unidecode==1.3
toml==0.10.2
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.27.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.1
wsproto==1.2.0
xmod==1.8.1
zope.interface==7.1.0
